api/requirements.txt
# === Core ===
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# === Deployment ===
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.2
psycopg2-binary>=2.9
python-decouple>=3.8
